external/vikash/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
anthropic==0.26.0
openai==1.30.1
PyPDF2==3.0.1
python-docx==1.1.2
redis==5.0.4
httpx==0.27.0
aiofiles==23.2.1
tiktoken==0.7.0
